삭제요청시  db,파일 삭제됨

+++ webtoon/views.py
@@ -1,3 +1,6 @@
+import os
+from django.conf import settings
+from django.http import HttpResponseForbidden
 from django.shortcuts import render,redirect,get_object_or_404
 from django.urls import reverse
 from .forms import WebtoonForm
@@ -32,11 +35,8 @@ def webtoon_create(request):
             'form': form
         }
         return render(request,'webtoon/create.html', context)
-
     elif request.method == 'POST':
         form = WebtoonForm(request.POST, request.FILES)
-        
-
         author = request.user
         title = request.POST['title']
         penname = request.POST['penname']
@@ -45,48 +45,41 @@ def webtoon_create(request):
         thumbnail = request.FILES['thumbnail']
         genre = request.POST['genre']
         tag = request.POST['tag'] 
-
         Webtoon.objects.create(author= author,title=title,penname=penname,description=description, image=image, thumbnail=thumbnail, genre=genre, tag=tag)
         return redirect('webtoon:webtoonmain')
     return render(request, 'create_webtoon.html')
-    
-
-
-    
+     
 @login_required
 def webtoon_update(request,pk):
     webtoon = get_object_or_404(Webtoon, pk=pk)
-
     if request.method == 'POST':
         form = WebtoonForm(request.POST, request.FILES, instance=webtoon,)
         if form.is_valid():
             form.save()
             return redirect('webtoon:webtoondetail', pk=pk)
-    else:
+    else:  
         form = WebtoonForm(instance=webtoon)
     print("수정안되쥬??")
     context = {
         'form': form,
         'webtoon': webtoon,
     }
-    
     return render(request, 'webtoon/update.html', context)
-
-
 
 @login_required
 def webtoon_delete(request, pk):
     webtoon = get_object_or_404(Webtoon, pk=pk)
+    if request.user != webtoon.author and not request.user.is_staff:
+        return HttpResponseForbidden("당신은 이 웹툰을 삭제할 권한이 없습니다.")
     if request.method == "POST":
         
+        webtoonpath =os.path.join(settings.MEDIA_ROOT,webtoon.image.path)
+        if os.path.exists(webtoonpath):
+            os.remove(webtoonpath)
+        thumbnailpath =os.path.join(settings.MEDIA_ROOT,webtoon.thumbnail.path)
+        os.remove(thumbnailpath)
         webtoon.delete()
         return redirect(reverse("webtoon:webtoonmain"))
     else:
         return redirect(reverse("webtoon:webtoondetail", kwargs={"pk": pk}))
     
-    
-
-
-
-
-
